refactor(messagebrokers): resolve reliability secondary facets by cast-at-consumption
Eliminate the AddReliabilityPair DI primitive and ReliabilityStoreLifetimeException.
The poll facet (IPollableOutboxStore) and inbox dedup facet are obtained by casting the
single resolved primary (IBrokeredMessageOutbox / IBrokeredMessageInbox) at the
consumption site, mirroring the existing IUnitOfWork cast. One registration per pair =
one resolved instance, so a split store is unrepresentable by construction; a custom
store lacking a facet throws InvalidCastException loudly at the poll site.
