+++ src/Chatter.MessageBrokers.Reliability.EntityFramework/src/Chatter.MessageBrokers.Reliability.EntityFramework/Extensions.cs
@@ -20,13 +20,12 @@ namespace Microsoft.Extensions.DependencyInjection
             return pipelineBuilder;
         }
 
-        public static CommandPipelineBuilder WithInboxBehavior<TContext>(this CommandPipelineBuilder pipelineBuilder) 
+        public static CommandPipelineBuilder WithInboxBehavior<TContext>(this CommandPipelineBuilder pipelineBuilder)
             where TContext : DbContext
         {
             pipelineBuilder.WithUnitOfWorkBehavior<TContext>();
             pipelineBuilder.Services.Replace<BrokeredMessageInbox<TContext>, BrokeredMessageInbox<TContext>>(ServiceLifetime.Scoped);
             pipelineBuilder.Services.Replace<IBrokeredMessageInbox>(ServiceLifetime.Scoped, sp => sp.GetRequiredService<BrokeredMessageInbox<TContext>>());
-            pipelineBuilder.Services.Replace<IInboxDeduplicator>(ServiceLifetime.Scoped, sp => sp.GetRequiredService<BrokeredMessageInbox<TContext>>());
             pipelineBuilder.WithBehavior(typeof(InboxBehavior<>));
 
             return pipelineBuilder;
@@ -36,11 +35,10 @@ namespace Microsoft.Extensions.DependencyInjection
             where TContext : DbContext
         {
             pipelineBuilder.WithBehavior(typeof(OutboxProcessingBehavior<>));
-            // Register the EF outbox once scoped and forward both the enqueue contract and the pollable store to
-            // the SAME instance so SendToOutbox and polling share one DbContext-bound store.
+            // Register the EF outbox once scoped and forward the enqueue contract to the same instance.
+            // IPollableOutboxStore is obtained by casting IBrokeredMessageOutbox at the consumption site.
             pipelineBuilder.Services.Replace<BrokeredMessageOutbox<TContext>, BrokeredMessageOutbox<TContext>>(ServiceLifetime.Scoped);
             pipelineBuilder.Services.Replace<IBrokeredMessageOutbox>(ServiceLifetime.Scoped, sp => sp.GetRequiredService<BrokeredMessageOutbox<TContext>>());
-            pipelineBuilder.Services.Replace<IPollableOutboxStore>(ServiceLifetime.Scoped, sp => sp.GetRequiredService<BrokeredMessageOutbox<TContext>>());
             pipelineBuilder.Services.Replace<IRouteBrokeredMessages, OutboxBrokeredMessageRouter>(ServiceLifetime.Scoped);
             pipelineBuilder.WithUnitOfWorkBehavior<TContext>();
