+++ src/Chatter.MessageBrokers/CONTEXT.md
@@ -54,7 +54,8 @@ _Avoid_: consumer, listener, middleware (honor existing `_Avoid_` aliases).
 **Inbox Deduplicator** (`IInboxDeduplicator`): the tier-neutral inbox dedup contract expressing once-only-handling intent (`HasBeenReceived`), implemented by both reliability tiers. Distinct from the relational-only `ReceiveViaInbox` wrap seam: this contract carries only the shared once-only intent, not the relational `Func<Task>` handler-wrap mechanics.
 _Avoid_: consumer, listener, middleware (honor existing `_Avoid_` aliases).
 
-**Reliability-Pair Lifetime Contract**: when a consumer supplies a single custom concrete satisfying both facets of a reliability pair (`IBrokeredMessageOutbox` + `IPollableOutboxStore`; `IBrokeredMessageInbox` + `IInboxDeduplicator`), it must register that concrete under the primary facet at `Scoped` or `Singleton` lifetime. The framework derives the secondary-facet registration at the same lifetime so both always resolve to the same instance (split-store is unrepresentable). A `Transient` custom primary is rejected at registration time because DI cannot guarantee same-instance resolution across two Transient resolutions. A consumer that registers both facets independently as separate descriptors pointing to different concretes owns ensuring the two registrations agree; the framework treats this as deliberate consumer intent and does not merge them.
+**Reliability-Store Facet Resolution**: the secondary reliability facet (`IPollableOutboxStore` for the outbox pair; `IInboxDeduplicator` for the inbox pair) is not an independent DI registration. It is obtained by casting the single resolved primary (`IBrokeredMessageOutbox` / `IBrokeredMessageInbox`) at the consumption site — the same precedent `OutboxProcessor` uses to obtain `IUnitOfWork` from the outbox. A custom primary that does not implement the required secondary facet throws `InvalidCastException` at the poll site (loud, not silent). Because there is exactly one resolved instance per pair, split-store is impossible by construction; no descriptor inspection, lifetime reconciliation, or fail-fast registration is required.
+_Avoid_: "Reliability-Pair Lifetime Contract" (the per-pair DI forwarding mechanism is eliminated).
 
 **ReceiveViaInbox / InboxBehavior**: the `ReceiveViaInbox(() => next())` wrap seam is relational-only mechanics. The inbox **contract** and intent (once-only dedup, idempotency) are shared across tiers; the seam that implements them is provider-specific. The document tier enlists the inbox marker via the Document-Tier Batch-Lifecycle Behavior and the document-tier reliability surface — not through `ReceiveViaInbox`.
