+++ src/Chatter.MessageBrokers/src/Chatter.MessageBrokers/Reliability/Outbox/IPollableOutboxStore.cs
@@ -11,14 +11,11 @@ namespace Chatter.MessageBrokers.Reliability.Outbox
     /// Outbox Relay.
     /// </summary>
     /// <remarks>
-    /// A custom reliability store that satisfies both <see cref="IBrokeredMessageOutbox"/> and
-    /// <see cref="IPollableOutboxStore"/> must be registered under <see cref="IBrokeredMessageOutbox"/> as a
-    /// single concrete at <c>Scoped</c> or <c>Singleton</c> lifetime. The framework forwards
-    /// <see cref="IPollableOutboxStore"/> to the same registration at the primary's lifetime so both facets
-    /// always resolve to the same instance. A <c>Transient</c> custom primary is rejected at registration time
-    /// because DI has no primitive to guarantee same-instance resolution across two Transient resolutions.
-    /// A consumer that registers both facets independently as separate descriptors owns ensuring the two
-    /// registrations agree; the framework does not merge them.
+    /// This is a secondary facet obtained by casting the single resolved <see cref="IBrokeredMessageOutbox"/>
+    /// at the consumption site — not an independent DI service. A custom reliability store must implement both
+    /// <see cref="IBrokeredMessageOutbox"/> and <see cref="IPollableOutboxStore"/> on one concrete registered
+    /// under <see cref="IBrokeredMessageOutbox"/>; a custom primary that does not implement this interface
+    /// throws <see cref="System.InvalidCastException"/> at the poll site.
     /// </remarks>
     public interface IPollableOutboxStore
     {
